Refactor announcements

+++ src/main/java/eu/daiad/web/domain/application/AlertByTypeRecord.java
@@ -6,20 +6,20 @@ public class AlertByTypeRecord
 {
     private final EnumAlertType type;
 
-    private final long count;
+    private final int count;
 
-    public AlertByTypeRecord(EnumAlertType type, long count)
+    public AlertByTypeRecord(EnumAlertType type, Long count)
     {
         this.type = type;
-        this.count = count;
+        this.count = count.intValue();
     }
 
-    public AlertByTypeRecord(String type, long count)
+    public AlertByTypeRecord(String type, Long count)
     {
         this(EnumAlertType.valueOf(type), count);
     }
 
-    public AlertByTypeRecord(int code, long count)
+    public AlertByTypeRecord(int code, Long count)
     {
         this(EnumAlertType.valueOf(code), count);
     }
@@ -29,7 +29,7 @@ public class AlertByTypeRecord
         return type;
     }
 
-    public long getCount()
+    public int getCount()
     {
         return count;
     }
